resumeGrader/requirements.txt
python-telegram-bot==20.6
openai
PyPDF2
python-dotenv
requests